WIP - Add async-operations typescript module
