+++ packages/async-operations/src/api.d.ts
@@ -1,0 +1,3 @@
+// This is to explicitly set model to 'any' as converting the model itself to TS
+// is out of scope.  PR reviewers - open to other options
+declare module '@cumulus/api/models';
