+++ packages/async-operations/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+      "outDir": "./dist"
+    },
+    "include": ["./src/*.ts"]
+  }
